changes in modified files are minor
mvaws.r and cmvaws.f contain the code to handle  independent mean (N(0,sigma^2))
and variance (sigma^2 * Chi^2 with  df degrees of freedom) estimates
This occures if both mean and variance are estimated from a linear model
First argument takes the estimated mean, sigma2 takes the estimated variance (error variance
in the linear model, not the variance of the mean !!!), df the degrees of freedom.
qlambda=.975 seems to work.


git-svn-id: svn+ssh://localhost/Home/stat/WorkSpace/aws-svn/trunk@32 78776fbd-6719-4c1f-9a4c-ebc1d401a228